add variable to vars
