+++ enclone_vars/src/export_code.rs
@@ -23,6 +23,7 @@ pub fn export_code(level: usize) -> Vec<(String, String)> {
         use amino::*;
         use crate::print_utils1::*;
         use crate::print_utils3::*;
+        use enclone_core::align_to_vdj_ref::*;
         use enclone_core::defs::*;
         use enclone_core::opt_d::*;
         use enclone_proto::types::*;
